// frontend/src/client/types.gen.ts
export type UserUpdate = {
  id: number
  name?: string | null
  role?: string | null
  status?: number | null
  supervisor_id?: number | null
  password?: string | null
}